Create code
The code is generic for whatever number of threads or barriers you use in the beginning. Which means, if I use N_THREADS = 5, it will print 5 stars first, then 5 + and the iteration will be similar all the way down (you can check yourself). The code will go through each barrier sequentially and once all barriers are used up it will start with the very first barrier and repeat until all the * and + have been printed. So, if I used N_BARRIERS = 4, my threads will be stopped by barrier 0 after they print star, then they will be stopped by barrier 1 after they print +. then they will be stopped by barrier 2 after they print * and so on till they reach barrier 5. If the for loop for printing still has not ended, barrier 0 will be reused and the whole process will repeat. 



Now, my code is very simple. I give two arguments to my barrier_wait => first one is the number of threads and the other is the barrier number I use. So, if I use barrier_wait(3, 0). It means I am using the function for the max number of threads of value 3 and I am using barrier number 0 to make all the three threads wait for synchronization.



I am using mutex to increment and decrement a counter. The counter keeps track of how many threads have entered and left a barrier. Since each thread will signal its previous thread before moving on, so in order to make sure my barrier is reusable I need to make sure that the last thread that leaves the barrier does not signal because there is no thread waiting before it. Hence, my barrier becomes reusable for the next iteration. I use the counter to keep track of when my last thread is about to leave and the if condition at the very end of my function shows this.

The rest of the code is pretty much self-explanatory!